django-boilerplate/requirements.txt
celery==5.4.0
cryptography==42.0.7
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
google-auth==2.29.0
gunicorn==22.0.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
